+++ CHANGELOG.md
@@ -5,8 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-[Unreleased]
+[2.1.0] 2026-02-25
 ### Changed
+- Single Volume view now show subtitles and Chapter Inde
 - Search Bar results has been enhanced.
 - Add support for symbol in footnotes anchors
 
@@ -353,7 +354,8 @@ the slug's name
 - Initial `index.php`, `style.css`, and `functions.php`
 - This Changelog
 
-[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v2.0.1...HEAD
+[Unreleased]: https://github.com/erremauro/cigno-zen/compare/v2.1.0...HEAD
+[2.1.0]: https://github.com/erremauro/cigno-zen/releases/tag/v2.1.0
 [2.0.1]: https://github.com/erremauro/cigno-zen/releases/tag/v2.0.1
 [2.0.0]: https://github.com/erremauro/cigno-zen/releases/tag/v2.0.0
 [1.13.0]: https://github.com/erremauro/cigno-zen/releases/tag/v1.13.0